Addresses few minor feedback: * Renames 'phone-in-hand.js' to 'phoneInHand.js' * Updates Celebrity Q & A tags with addition of 'Interview'. * Adds space for Yellow Heart sticker lable.
