+++ packages/stickers/src/how-contact-tracing-works/phoneInHand.js
