+++ packages/stickers/src/summer-fashion-collection/yellowHeart.js
@@ -19,7 +19,7 @@
 import { _x } from '@web-stories-wp/i18n';
 import PropTypes from 'prop-types';
 
-const title = _x('YellowHeart', 'sticker name', 'web-stories');
+const title = _x('Yellow Heart', 'sticker name', 'web-stories');
 
 const YellowHeart = ({ style }) => (
   <svg
